+++ index.html
@@ -227,7 +227,7 @@
     </section>
   </template>
 
-  <script src="js/functions.js"></script>
+  <script src="js/main.js" type="module"></script>
 
 </body>
 </html>
